% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{kl_divergence_metric}
\alias{kl_divergence_metric}
\title{Calculate Mean Kullback-Leibler (KL) Divergence Metric}
\usage{
kl_divergence_metric(true_y, pred_y, smoothing_value = 1e-09)
}
\arguments{
\item{true_y}{A numeric matrix of non-negative true target values.}

\item{pred_y}{A numeric matrix of non-negative predicted target values.}

\item{smoothing_value}{A small positive constant added to all values before
normalization. Defaults to `1e-9`.}
}
\value{
A single numeric value representing the mean KL divergence.
}
\description{
Calculates the Kullback-Leibler (KL) divergence between true
  and predicted probability distributions on a row-wise basis and returns the
  mean divergence. This function is designed to be used as a custom
  `metric_function` in analyses like `permutation_feature_importance`.
}
\details{
This function treats each row of the input matrices (`true_y`, `pred_y`)
  as a separate probability distribution. For each row, it calculates the
  KL divergence DKL(P || Q) and returns the mean of these values.
}
